refact folder and objects
