+++ package.json
@@ -4,8 +4,10 @@
   "description": "Nice :D",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack --mode=development"
+    "test": "jest",
+    "watch": "jest --watch",
+    "build": "webpack --mode=development",
+    "dev": "webpack-dev-server --open"
   },
   "author": "",
   "license": "ISC",
@@ -18,6 +20,7 @@
   },
   "homepage": "https://github.com/SotoArmando/odin-battleships#readme",
   "dependencies": {
+    "jest": "^25.1.0",
     "webpack": "^4.41.6"
   },
   "devDependencies": {
